+++ changelog.md
@@ -1,3 +1,4 @@
 ## General
 
+- Added carpet rule fakePlayerForceOffline
 - Added support for Minecraft 1.21.5 and 1.21.6